Links between Assays

Links between assays within a Features object are handled by an AssayLinks object. It is composed by a list of AssayLink instances.

# S4 method for AssayLink
show(object)

AssayLink(name, from = NA_character_, fcol = NA_character_, hits = Hits())

AssayLinks(..., names = NULL)

assayLink(x, i)

assayLinks(x, i)

# S4 method for AssayLink,character,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for AssayLinks,list,ANY,ANY
[(x, i, j, ..., drop = TRUE)

Arguments

object
An AssayLink object to show.

name
A mandatory name of the assay(s).

from
The name of the parent assay, or NA_character_, if not applicable.

fcol
The feature variable of the parent assay used to generate the current assay (used in aggregateFeatures). NA_character_, if not applicable.

hits
An object of class S4Vectors::Hits matching the features of two assays.

...
A set of AssayLink objects or a list thereof.

names
A character() of AssayLink names. If provided, ... are ignored, and names is used to create an AssayLinks object with AssayLink instances with names names.

x
An instance of class Features.

i
The index or name of the assay whose AssayLink and parents AssayLink instances are to be returned. For [, the feature names to filter on.

j
ignored.

drop
ignored.

Value

assayLink returns an instance of class AssayLink.

assayLinks returns an instance of class AssayLinks.

Constructors

Object can be created with the AssayLink() and AssayLinks() constructors.

Methods and functions

assayLink(x, i) accesses the AssayLink at position i or with name i in the Features object x.

parentAssayLinks(x, i, recursive = FALSE) accesses the parent(s) AssayLinks or assay with index or name i.

Examples

al1 <- AssayLink(name = "assay1")
al1

#> AssayLink for assay <assay1>
#> [from:NA|fcol:NA|hits:0]

AssayLinks(al1)

#> AssayLinks of length 1
#> names(1): assay1

al2 <- AssayLinks(names = c("Assay1", "Assay2"))
al2

#> AssayLinks of length 2
#> names(2): Assay1 Assay2