Simlple container for an assay and its featre metadata

A FeatureSet is a simple container to store quantitative data (called the assay) and the associated feature metadata (called the featureData). The number of rows and the rownames of the assay and feature metadata must match exactly.

FeatureSet(assay = matrix(ncol = 0, nrow = 0),
  featureData = DataFrame())

# S4 method for FeatureSet
sampleNames(object)

# S4 method for FeatureSet,character
sampleNames(object) <- value

# S4 method for FeatureSet
dim(x)

# S4 method for FeatureSet
ncol(x)

# S4 method for FeatureSet
nrow(x)

# S4 method for FeatureSet,ANY
assay(x, i, ...)

# S4 method for FeatureSet,missing
featureVariables(x, i, ...)

# S4 method for FeatureSet,missing
featureData(x, i, ...)

# S4 method for FeatureSet,missing
featureNames(x, i, ...)

# S4 method for FeatureSet,ANY,ANY,missing
[(x, i, j, ..., drop = FALSE)

Arguments

assay
A `matrix` containing the quantitation data.

featureData
A `DataFrame` with feature annotations.

object
A `FeatureSet` object.

value
An appropriate replacement value.

x
The `FeatureSet` to subset.

i
Subsetting vector for the assay and featureData's rows.

...
Additional paramaeters (ignored).

j
Subsetting vector for the assay's columns.

drop
Always set to `FALSE`.

Slots

assay

A matrix of numerics that stores quantitative data for a set of features (along the rows) and samples (along the columns).

featureData

A DataFrame to hold the features annotations.

links

A object of class FeatureSetLinks holding the information tracing back to the object's parent FeatureSet. For internal use at the Features-level only.

version

A character(1) providing the class version. For internal use only.

See also

Examples

## An empty FeatureSet
FeatureSet()

#> class: FeatureSet 
#> dim: 0 0 
#> Features(0):
#> Samples(0):
#> Feature variables(0):

## See ?Features for more examples