Quantitative MS Features

Conceptually, a Features object holds a set of matrix (or array) elements containing quantitative data. The number of columns (samples) are always the same across the matrices, but the number of rows (features) can vary. Each one of these matrices has a set of feature annotations (encoded as DataFrame objects), that have the same number of rows as the assay matrix their are associated to, and an arbitrary number of columns (feature variables). Such a matrix and feature annotation pair is called a FeatureSet. In addition, a Features object also uses a single DataFrame to annotate the samples (columns) represented in all the matrices.

# S4 method for Features
dims(x)

# S4 method for Features
dim(x)

# S4 method for Features
sampleNames(object)

# S4 method for Features,character
sampleNames(object) <- value

# S4 method for Features
colData(x)

# S4 method for Features,DataFrame
colData(x) <- value

# S4 method for Features,missing
featureVariables(x, i, ...)

# S4 method for Features,numeric
featureVariables(x, i, ...)

# S4 method for Features,character
featureVariables(x, i, ...)

# S4 method for Features,missing
features(x, i, ...)

# S4 method for Features
assays(x, ..., withDimnames = TRUE)

# S4 method for Features,numeric
assay(x, i, ...)

# S4 method for Features,character
assay(x, i, ...)

# S4 method for Features,missing
assay(x, i, ...)

# S4 method for Features,missing
featureData(x, i, ...)

# S4 method for Features,numeric
featureData(x, i, ...)

# S4 method for Features,character
featureData(x, i, ...)

# S4 method for Features,missing
featureNames(x, i, ...)

# S4 method for Features,numeric
featureNames(x, i, ...)

# S4 method for Features,character
featureNames(x, i, ...)

Features(..., colData = NULL, metadata = list())

Arguments

x
An instance of class Features.

object
An instance of class Features.

value
The replacement value.

i
A character(1) or numeric(1) for subsetting.

...
One or multiple [FeatureSet] instances passed to the `Features` constructor.

withDimnames
Always `TRUE`.

colData
A `DataFrame` with column (sample annotations).

metadata
A `list()` with arbitrary object annotations.

Value

A new instance of class `Features`.

Details

A typical use case for such Features object is to represent quantitative proteomics (or metabolomics) data, where different assays represent quantitation data at the PSM (the main assay), peptide and protein level, and where peptide values are computed from the PSM data, and the protein-level data is calculated based on the peptide-level values.

The largest FeatureSet (the one with the highest number of features) is considered the main assay, from which the other ones are derived by aggregating/combining several rows into a single one.

The recommended way to create Features objects is the use the readFeatures() function, that creates an instance from tabular data. The Features constructor can be used to create objects from their bare parts. It is the user's responsability to make sure that these match the class validity requirements.

A Features instance must comply with the following requirements:

The features are stored as FeatureSet instances, and each FeatureSet has the same number of columns.

The number of rows and the names of colData match the number and names of the samples (columns) in the FeatureSets.

General metadata about the object itself is provided in the metadata list, which can be empty.

In the usage section, the respective arguments correspond to:

Accessors

The following accessors, inheritied from SimpleList are available:

length(x): returns the number of FeatureSet instances in x's FeatureSets.

names(x), names(x) <- value: gets or sets the assays optional names of the FeatureSet instances. value is a character of length equal to length(x).

dims(x): returns the dimensions of all FeatureSet's assays.

dim(x): returns the dimensions of the largest assay.

isEmpty(x): returns TRUE for an object without any FeatureSets, FALSE otherwise.

metadata(x), metadata(x) <- value: gets and sets the object's global metadata. value must be a list.

x[[i]], x[[i]] <- value: gets and sets a single FeatureSet element, where i is either a numeric(1) or a character(1). value must be of class FeatureSet.

x[i], x[i] <- value: gets the list of FeatureSet instances defined the numeric or character i. Can also be used to replace the the subset of FeatureSet elements with a new Features values of matching samples.

The following accessors are also available.

sampleNames(object), sampleNames(x) <- value: gets and sets the samples names. value must be a character of appropriate length.

colData(x), colData(x) <- value: gets or sets the columns/samples metadata.value must be a DataFrame object. Row names of value match the existing column names of the assays.

featureVariables(x): gets the list of feature variables, where each element of the list is a character of names for the corresponding FeatureSet.

featureVariables(object, i): a convenient alternative (to featureVariables(x)[[i]]) to get the ith featureData element.

assays(x): gets a list of assays.

assay(x, i)(x, i): A convenient alternative (to assays(x)[[i]]) to get the ith (default is the largest) assay element.

featureData(x): gets all assays' feature metadata as a list.

featureData(x, i): a convenient alternative (to featureData(x)[[i]]) to get the ith featureData element.

featureNames(x): gets the list of feature names, where each element of the list is a character of names for the corresponding assay.

featureNames(object, i): a convenient alternative (to featureNames(x)[[i]]) to get the ith featureData element.

See also

Examples

## ------------------------
## An empty Features object
## ------------------------

Features()

#> class: Features 
#> metadata(0):
#> FeatureSets(0):
#> Samples(0):
#> colData(0):

## -----------------------------------
## Creating a Features object manually
## -----------------------------------

## two assays (matrices) with matching column names
m1 <- matrix(1:40, ncol = 4)
m2 <- matrix(1:16, ncol = 4)
sample_names <- paste0("S", 1:4)
colnames(m1) <- colnames(m2) <- sample_names
rownames(m1) <- letters[1:10]
rownames(m2) <- letters[23:26]

## two corresponding feature metadata with appropriate row names
df1 <- DataFrame(Fa = 1:10, Fb = letters[1:10],
                 row.names = rownames(m1))
df2 <- DataFrame(row.names = rownames(m2))

## two FeatureSets, that combine the respecitve assays and feature
## metadata
fs1 <- FeatureSet(m1, df1)
fs1

#> class: FeatureSet 
#> dim: 10 4 
#> Features(10): a b ... i j
#> Samples(4): S1 S2 S3 S4
#> Feature variables(2): Fa Fb

fs2 <- FeatureSet(m2, df2)
fs2

#> class: FeatureSet 
#> dim: 4 4 
#> Features(4): w x y z
#> Samples(4): S1 S2 S3 S4
#> Feature variables(0):

## Sample annotation (colData)
cd <- DataFrame(Var1 = rnorm(4),
                Var2 = LETTERS[1:4],
                row.names = sample_names)

## Putting the FeatureSets and their common sample annoation
## together into a Features object
fts1 <- Features(fs1, fs2,
                 colData = cd)
fts1

#> class: Features 
#> metadata(0):
#> FeatureSets(2): '' ''
#> Samples(4): S1 S2 S3 S4
#> colData(2): Var1 Var2

## -------------------------------------------------
## Creating a Features object from a data.frame (see
## ?readFeatures) for details
## -------------------------------------------------

data(hlpsms)
fts2 <- readFeatures(hlpsms, ecol = 1:10, name = "psms")
fts2

#> class: Features 
#> metadata(0):
#> FeatureSets(1): psms
#> Samples(10): X126 X127C ... X130N X131
#> colData(0):

features(fts2)

#> $psms
#> class: FeatureSet 
#> dim: 32776 10 
#> Features(32776): 136 141 ... 115747 115775
#> Samples(10): X126 X127C ... X130N X131
#> Feature variables(18): Sequence ProteinDescriptions ... RTmin markers
#> 

fts2[[1]]

#> class: FeatureSet 
#> dim: 32776 10 
#> Features(32776): 136 141 ... 115747 115775
#> Samples(10): X126 X127C ... X130N X131
#> Feature variables(18): Sequence ProteinDescriptions ... RTmin markers

fts2[["psms"]]

#> class: FeatureSet 
#> dim: 32776 10 
#> Features(32776): 136 141 ... 115747 115775
#> Samples(10): X126 X127C ... X130N X131
#> Feature variables(18): Sequence ProteinDescriptions ... RTmin markers