Filter features based on their rowData

The filterFeatures methods enables users to filter features based on a variable in their rowData. The features matching the filter will be returned as a new object of class Features. The filters can be provided as instances of class AnnotationFilter (see below) or as formulas.

VariableFilter(field, value, condition = "==")

# S4 method for Features,AnnotationFilter
filterFeatures(object, filter, ...)

# S4 method for Features,formula
filterFeatures(object, filter, ...)

Arguments

field
character(1) refering to the name of the variable to apply the filter on.

value
‘character()’ or ‘integer()’ value for the CharacterVariableFilter and NumericVariableFilter filters respectively.

condition
‘character(1)’ defining the condition to be used in the filter. For ‘NumericVariableFilter’, one of ‘"=="’, ‘"!="’, ‘">"’, ‘"<"’, ‘">="’ or ‘"<="’. For ‘CharacterVariableFilter’, one of ‘"=="’, ‘"!="’, ‘"startsWith"’, ‘"endsWith"’ or ‘"contains"’. Default condition is ‘"=="’.

object
An instance of class Features.

filter
Either an instance of class AnnotationFilter or a formula.

...
Additional parameters. Currently ignored.

Variable filters

The variable filters are filters as defined in the AnnotationFilter package. In addition to the pre-defined filter, users can arbitrarily set a field on which to operate. These arbitrary filters operate either on a character variables (as CharacterVariableFilter objects) or numerics (as NumericVariableFilters objects), which can be created with the VariableFilter constructor.

See also

Examples

## ----------------------------------------
## Creating character and numberic
## variable filters
## ----------------------------------------

VariableFilter(field = "my_var",
               value = "value_to_keep",
               condition = "==")

#> class: CharacterVariableFilter 
#> condition: == 
#> value: value_to_keep 

VariableFilter(field = "my_num_var",
               value = 0.05,
               condition = "<=")

#> class: NumericVariableFilter 
#> condition: <= 
#> value: 0.05 

example(aggregateFeatures)

#> 
#> aggrgF> ## An example Features with PSM-level data
#> aggrgF> data(feat1)
#> 
#> aggrgF> ## Aggregate PSMs into peptides
#> aggrgF> feat1 <- aggregateFeatures(feat1, "psms", "Sequence", name = "peptides")
#> 
#> aggrgF> feat1
#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 10 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 3 rows and 2 columns 
#> 
#> aggrgF> ## Aggregate peptides into proteins
#> aggrgF> feat1 <- aggregateFeatures(feat1, "peptides", "Protein", name = "proteins")
#> 
#> aggrgF> feat1
#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 10 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 3 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 2 rows and 2 columns 
#> 
#> aggrgF> ## Aggregation with missing values
#> aggrgF> data(ft_na)
#> 
#> aggrgF> ft_na
#> A instance of class Features containing
#>  [1] na: SummarizedExperiment with 4 rows and 3 columns 
#> 
#> aggrgF> assay(ft_na, 1)
#>    A  B  C
#> a NA  5  9
#> b  2  6 10
#> c  3 NA 11
#> d NA  8 12
#> 
#> aggrgF> rowData(ft_na[[1]])
#> DataFrame with 4 rows and 1 column
#>           X
#>   <integer>
#> a         1
#> b         2
#> c         1
#> d         2
#> 
#> aggrgF> ## By default, missing values are propagated
#> aggrgF> assay(aggregateFeatures(ft_na, 1, fcol = "X", fun = colSums), 2)

#> Your data contains missing values. Please read the relevant section in
#> the aggregateFeatures manual page for details the effects of missing
#> values on data aggregation.

#>    A  B  C
#> 1 NA NA 20
#> 2 NA 14 22
#> 
#> aggrgF> ## Ignored when setting na.rm = TRUE
#> aggrgF> assay(aggregateFeatures(ft_na, 1, fcol = "X", fun = colSums, na.rm = TRUE), 2)

#> Your data contains missing values. Please read the relevant section in
#> the aggregateFeatures manual page for details the effects of missing
#> values on data aggregation.

#>   A  B  C
#> 1 3  5 20
#> 2 2 14 22

## ----------------------------------------------------------------
## Filter all features that are associated to the Mitochondrion in
## the location feature variable. This variable is present in all
## assays.
## ----------------------------------------------------------------

## using the forumla interface, exact mathc
filterFeatures(feat1, ~  location == "Mitochondrion")

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 6 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 2 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 1 rows and 2 columns 

## using the forumula intefrace, martial match
filterFeatures(feat1, ~startsWith(location, "Mito"))

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 6 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 2 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 1 rows and 2 columns 

## using a user-defined character filter
filterFeatures(feat1, VariableFilter("location", "Mitochondrion"))

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 6 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 2 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 1 rows and 2 columns 

## ----------------------------------------------------------------
## Filter all features that aren't marked as unknown (sub-cellular
## location) in the feature variable
## ----------------------------------------------------------------

## using a user-defined character filter
filterFeatures(feat1, VariableFilter("location", "unknown", condition = "!="))

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 6 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 2 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 1 rows and 2 columns 

## using the forumula interface
filterFeatures(feat1, ~ location != "unknown")

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 6 rows and 2 columns 
#>  [2] peptides: SummarizedExperiment with 2 rows and 2 columns 
#>  [3] proteins: SummarizedExperiment with 1 rows and 2 columns 

## ----------------------------------------------------------------
## Filter features that have a p-values lower or equal to 0.03
## ----------------------------------------------------------------

## using a user-defined numeric filter
filterFeatures(feat1, VariableFilter("pval", 0.03, "<="))

#> harmonizing input:
#>   removing 4 sampleMap rows not in names(experiments)

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 5 rows and 2 columns 

## using the formula interface
filterFeatures(feat1, ~ pval <= 0.03)

#> harmonizing input:
#>   removing 4 sampleMap rows not in names(experiments)

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 5 rows and 2 columns 

## ----------------------------------------------------------------
## Negative control - filtering for an non-existing markers value
## or a missing feature variable, returning empty results
## ----------------------------------------------------------------

filterFeatures(feat1, VariableFilter("location", "not"))

#> A empty instance of class Features 

filterFeatures(feat1, ~ location == "not")

#> A empty instance of class Features 

filterFeatures(feat1, VariableFilter("foo", "bar"))

#> A empty instance of class Features 

filterFeatures(feat1, ~ foo == "bar")

#> A empty instance of class Features