Managing missing data

This manual page describes the handling of missing values in Features objects. In the following functions, if object is of class Features, and optional assay index or name i can be specified to define the assay (by name of index) on which to operate.

The following functions are currently available:

zeroIsNA(object, i) replaces all 0 in object by NA. This is often necessary when third-party software assume that features that weren't quantified should be assigned an intensity of 0.

nNA(object, i) return a list of missing value summaries. The first element nNA gives the percentage of missing values; the second element nNArows provides a table of the number of missing values for the features (rows) of the assay(s); the third element nNAcols provides the number of missing values in each sample of the assay(s).

filterNA(object, pNA, i) removes features (rows) that contain pNA percentage or more missing values.

See the Processing vignette for examples.

# S4 method for SummarizedExperiment,missing
zeroIsNA(object, i)

# S4 method for Features,integer
zeroIsNA(object, i)

# S4 method for Features,numeric
zeroIsNA(object, i)

# S4 method for Features,character
zeroIsNA(object, i)

# S4 method for SummarizedExperiment,missing
nNA(object, i)

# S4 method for Features,integer
nNA(object, i)

# S4 method for Features,numeric
nNA(object, i)

# S4 method for Features,character
nNA(object, i)

# S4 method for SummarizedExperiment
filterNA(object, pNA = 0)

# S4 method for Features
filterNA(object, pNA = 0, i)

Arguments

object
An object of class Features or SummarizedExperiment.

i
One or more indices or names of the assay(s) to be processed.

pNA
numeric(1) providing the maximim percentage of missing values per feature (row) that is acceptable. Feature with higher percentages are removed. If 0 (default), features that contain any number of NA values are dropped.

Value

An instance of the same class as object.