Reduces and expands a `DataFrame`

A long dataframe can be *reduced* by mergeing certain rows into a single one. These new variables are constructed as a `SimpleList` containing all the original values. Invariant columns, i.e columns that have the same value along all the rows that need to be merged, can be shrunk into a new variables containing that invariant value (rather than in list columns). The grouping of rows, i.e. the rows that need to be shrunk together as one, is defined by a vector.

reduceDataFrame(x, k, count = FALSE, simplify = TRUE, drop = FALSE)

expandDataFrame(x, k = NULL)

Arguments

x
The `DataFrame` to be reduced or expanded.

k
A ‘vector’ of length `nrow(x)` defining the grouping based on which the `DataFrame` will be shrunk.

count
`logical(1)` specifying of an additional column (called by default `.n`) with the tally of rows shrunk into on new row should be added. Note that if already existing, `.n` will be silently overwritten.

simplify
A `logical(1)` defining if invariant columns should be converted to simple lists. Default is `TRUE`.

drop
A `logical(1)` specifying whether the non-invariant columns should be dropped altogether. Default is `FALSE`.

Value

An expanded (reduced) `DataFrame`.

Details

The opposite operation is *expand*. But note that for a `DataFrame` to be expanded back, it must not to be simplified.

Examples

library("IRanges")

k <- sample(100, 1e3, replace = TRUE)
df <- DataFrame(k = k,
                x = round(rnorm(length(k)), 2),
                y = seq_len(length(k)),
                z = sample(LETTERS, length(k), replace = TRUE),
                ir = IRanges(seq_along(k), width = 10),
                r = Rle(sample(5, length(k), replace = TRUE)),
                invar = k + 1)
df

#> DataFrame with 1000 rows and 7 columns
#>              k         x         y           z        ir     r     invar
#>      <integer> <numeric> <integer> <character> <IRanges> <Rle> <numeric>
#> 1           31      -0.5         1           U      1-10     2        32
#> 2           68      0.95         2           H      2-11     1        69
#> 3           73     -0.17         3           G      3-12     3        74
#> 4           69     -1.11         4           U      4-13     5        70
#> 5            5     -0.95         5           A      5-14     4         6
#> ...        ...       ...       ...         ...       ...   ...       ...
#> 996         10      0.08       996           N  996-1005     5        11
#> 997         56      0.56       997           J  997-1006     5        57
#> 998         98     -0.17       998           H  998-1007     1        99
#> 999          2     -0.23       999           Q  999-1008     1         3
#> 1000        32      0.24      1000           M 1000-1009     4        33

## Shinks the DataFrame
df2 <- reduceDataFrame(df, df$k)
df2

#> DataFrame with 100 rows and 7 columns
#>             k                     x               y               z
#>     <integer>         <NumericList>   <IntegerList> <CharacterList>
#> 1           1    2.48,1.5,-0.61,...  42,381,492,...       Y,H,Q,...
#> 2           2 -1.14,-1.02,-2.02,...     9,60,69,...       J,S,A,...
#> 3           3 -0.77,-1.53,-0.23,...  62,130,191,...       M,E,F,...
#> 4           4  0.41,-0.68,-0.44,...  22,114,155,...       R,O,A,...
#> 5           5   -0.95,0.61,0.14,...     5,15,84,...       A,V,Q,...
#> ...       ...                   ...             ...             ...
#> 96         96   -1.64,0.46,2.03,...  52,327,444,...       A,B,I,...
#> 97         97  -0.38,1.22,-0.28,...   74,88,323,...       A,B,P,...
#> 98         98    0.46,1.68,-0.1,... 232,274,306,...       V,H,I,...
#> 99         99   0.92,1.45,-0.86,... 397,399,474,...       T,N,W,...
#> 100       100   0.16,-0.96,0.25,... 128,149,255,...       U,J,X,...
#>                              ir         r     invar
#>                   <IRangesList> <RleList> <numeric>
#> 1     42-51,381-390,492-501,... 5,4,2,...         2
#> 2          9-18,60-69,69-78,... 2,5,1,...         3
#> 3     62-71,130-139,191-200,... 2,2,3,...         4
#> 4     22-31,114-123,155-164,... 3,1,3,...         5
#> 5          5-14,15-24,84-93,... 4,5,2,...         6
#> ...                         ...       ...       ...
#> 96    52-61,327-336,444-453,... 1,5,2,...        97
#> 97      74-83,88-97,323-332,... 3,1,1,...        98
#> 98  232-241,274-283,306-315,... 4,5,5,...        99
#> 99  397-406,399-408,474-483,... 2,5,4,...       100
#> 100 128-137,149-158,255-264,... 4,1,1,...       101

## With a tally of the number of members in each group
reduceDataFrame(df, df$k, count = TRUE)

#> DataFrame with 100 rows and 8 columns
#>             k                     x               y               z
#>     <integer>         <NumericList>   <IntegerList> <CharacterList>
#> 1           1    2.48,1.5,-0.61,...  42,381,492,...       Y,H,Q,...
#> 2           2 -1.14,-1.02,-2.02,...     9,60,69,...       J,S,A,...
#> 3           3 -0.77,-1.53,-0.23,...  62,130,191,...       M,E,F,...
#> 4           4  0.41,-0.68,-0.44,...  22,114,155,...       R,O,A,...
#> 5           5   -0.95,0.61,0.14,...     5,15,84,...       A,V,Q,...
#> ...       ...                   ...             ...             ...
#> 96         96   -1.64,0.46,2.03,...  52,327,444,...       A,B,I,...
#> 97         97  -0.38,1.22,-0.28,...   74,88,323,...       A,B,P,...
#> 98         98    0.46,1.68,-0.1,... 232,274,306,...       V,H,I,...
#> 99         99   0.92,1.45,-0.86,... 397,399,474,...       T,N,W,...
#> 100       100   0.16,-0.96,0.25,... 128,149,255,...       U,J,X,...
#>                              ir         r     invar        .n
#>                   <IRangesList> <RleList> <numeric> <integer>
#> 1     42-51,381-390,492-501,... 5,4,2,...         2        10
#> 2          9-18,60-69,69-78,... 2,5,1,...         3        10
#> 3     62-71,130-139,191-200,... 2,2,3,...         4         9
#> 4     22-31,114-123,155-164,... 3,1,3,...         5        11
#> 5          5-14,15-24,84-93,... 4,5,2,...         6        12
#> ...                         ...       ...       ...       ...
#> 96    52-61,327-336,444-453,... 1,5,2,...        97         8
#> 97      74-83,88-97,323-332,... 3,1,1,...        98        12
#> 98  232-241,274-283,306-315,... 4,5,5,...        99         8
#> 99  397-406,399-408,474-483,... 2,5,4,...       100        12
#> 100 128-137,149-158,255-264,... 4,1,1,...       101        10

## Much faster, but more crowded result
df3 <- reduceDataFrame(df, df$k, simplify = FALSE)
df3

#> DataFrame with 100 rows and 7 columns
#>                   k                     x               y               z
#>       <IntegerList>         <NumericList>   <IntegerList> <CharacterList>
#> 1         1,1,1,...    2.48,1.5,-0.61,...  42,381,492,...       Y,H,Q,...
#> 2         2,2,2,... -1.14,-1.02,-2.02,...     9,60,69,...       J,S,A,...
#> 3         3,3,3,... -0.77,-1.53,-0.23,...  62,130,191,...       M,E,F,...
#> 4         4,4,4,...  0.41,-0.68,-0.44,...  22,114,155,...       R,O,A,...
#> 5         5,5,5,...   -0.95,0.61,0.14,...     5,15,84,...       A,V,Q,...
#> ...             ...                   ...             ...             ...
#> 96     96,96,96,...   -1.64,0.46,2.03,...  52,327,444,...       A,B,I,...
#> 97     97,97,97,...  -0.38,1.22,-0.28,...   74,88,323,...       A,B,P,...
#> 98     98,98,98,...    0.46,1.68,-0.1,... 232,274,306,...       V,H,I,...
#> 99     99,99,99,...   0.92,1.45,-0.86,... 397,399,474,...       T,N,W,...
#> 100 100,100,100,...   0.16,-0.96,0.25,... 128,149,255,...       U,J,X,...
#>                              ir         r           invar
#>                   <IRangesList> <RleList>   <NumericList>
#> 1     42-51,381-390,492-501,... 5,4,2,...       2,2,2,...
#> 2          9-18,60-69,69-78,... 2,5,1,...       3,3,3,...
#> 3     62-71,130-139,191-200,... 2,2,3,...       4,4,4,...
#> 4     22-31,114-123,155-164,... 3,1,3,...       5,5,5,...
#> 5          5-14,15-24,84-93,... 4,5,2,...       6,6,6,...
#> ...                         ...       ...             ...
#> 96    52-61,327-336,444-453,... 1,5,2,...    97,97,97,...
#> 97      74-83,88-97,323-332,... 3,1,1,...    98,98,98,...
#> 98  232-241,274-283,306-315,... 4,5,5,...    99,99,99,...
#> 99  397-406,399-408,474-483,... 2,5,4,... 100,100,100,...
#> 100 128-137,149-158,255-264,... 4,1,1,... 101,101,101,...

## Drop all non-invariant columns
reduceDataFrame(df, df$k, drop = TRUE)

#> DataFrame with 100 rows and 2 columns
#>             k     invar
#>     <integer> <numeric>
#> 1           1         2
#> 2           2         3
#> 3           3         4
#> 4           4         5
#> 5           5         6
#> ...       ...       ...
#> 96         96        97
#> 97         97        98
#> 98         98        99
#> 99         99       100
#> 100       100       101