Reduces and expands a `DataFrame`

A long dataframe can be *reduced* by mergeing certain rows into a single one. These new variables are constructed as a `SimpleList` containing all the original values. Invariant columns, i.e columns that have the same value along all the rows that need to be merged, can be shrunk into a new variables containing that invariant value (rather than in list columns). The grouping of rows, i.e. the rows that need to be shrunk together as one, is defined by a vector.

reduce_DataFrame(x, k, count = FALSE, simplify = TRUE)

expand_DataFrame(x, k = NULL)

Arguments

x
The `DataFrame` to be reduced or expanded.

k
A ‘vector’ of length `nrow(x)` defining the grouping based on which the `DataFrame` will be shrunk.

count
`logical(1)` specifying of an additional column (called by default `.n`) with the tally of rows shrunk into on new row should be added. Note that if already existing, `.n` will be silently overwritten.

simplify
`logical(1)` defining if invariant columns should be converted to simple lists.

Value

An expanded (reduced) `DataFrame`.

Details

The opposite operation is *expand*. But note that for a `DataFrame` to be expanded back, it must not to be simplified.

Examples

library("IRanges")

k <- sample(10000, 1e5, replace = TRUE)
df <- DataFrame(k = k,
                x = round(rnorm(length(k)), 2),
                y = seq_len(length(k)),
                z = sample(LETTERS, length(k), replace = TRUE),
                ir = IRanges(seq_along(k), width = 10),
                r = Rle(sample(5, length(k), replace = TRUE)),
                invar = k + 1)
df

#> DataFrame with 100000 rows and 7 columns
#>                k         x         y           z            ir     r     invar
#>        <integer> <numeric> <integer> <character>     <IRanges> <Rle> <numeric>
#> 1           1476     -1.52         1           X          1-10     5      1477
#> 2           8166     -0.07         2           H          2-11     3      8167
#> 3           2245     -0.23         3           D          3-12     5      2246
#> 4           4613      1.09         4           K          4-13     2      4614
#> 5           9983      0.02         5           W          5-14     2      9984
#> ...          ...       ...       ...         ...           ...   ...       ...
#> 99996       1454      0.28     99996           I  99996-100005     1      1455
#> 99997       6950     -2.93     99997           L  99997-100006     2      6951
#> 99998       3870      0.54     99998           G  99998-100007     2      3871
#> 99999       7972      0.05     99999           A  99999-100008     4      7973
#> 100000      3536      1.09    100000           Y 100000-100009     1      3537

## Shinks the DataFrame
df2 <- reduce_DataFrame(df, df$k)
df2

#> DataFrame with 10000 rows and 7 columns
#>               k                     x                     y               z
#>       <integer>         <NumericList>         <IntegerList> <CharacterList>
#> 1             1 -0.65,-0.11,-0.53,... 36836,43803,45720,...       J,N,O,...
#> 2             2    0.06,0.16,0.13,...   6865,7895,32259,...       W,N,Y,...
#> 3             3     -1.4,0.34,1.3,...  5416,10697,16559,...       K,R,T,...
#> 4             4    0.85,0.69,0.96,...  6721,16775,19061,...       P,I,C,...
#> 5             5   -1.1,0.04,-0.29,...  7463,12833,16519,...       N,G,M,...
#> ...         ...                   ...                   ...             ...
#> 9996       9996   -1.81,0.19,0.37,...    887,9897,12215,...       O,G,I,...
#> 9997       9997   0.98,-1.66,0.07,...   3126,4474,12738,...       A,M,J,...
#> 9998       9998    -0.9,0.04,0.62,... 13651,14503,16927,...       L,B,W,...
#> 9999       9999   1.73,0.84,-0.08,...  1940,20097,27228,...       J,M,L,...
#> 10000     10000   -0.91,0.56,0.45,... 10600,21517,40461,...       X,W,I,...
#>                                            ir         r     invar
#>                                 <IRangesList> <RleList> <numeric>
#> 1     36836-36845,43803-43812,45720-45729,... 3,2,3,...         2
#> 2         6865-6874,7895-7904,32259-32268,... 5,2,1,...         3
#> 3       5416-5425,10697-10706,16559-16568,... 5,2,4,...         4
#> 4       6721-6730,16775-16784,19061-19070,... 5,5,4,...         5
#> 5       7463-7472,12833-12842,16519-16528,... 1,1,1,...         6
#> ...                                       ...       ...       ...
#> 9996        887-896,9897-9906,12215-12224,... 4,3,2,...      9997
#> 9997      3126-3135,4474-4483,12738-12747,... 1,4,1,...      9998
#> 9998  13651-13660,14503-14512,16927-16936,... 4,5,1,...      9999
#> 9999    1940-1949,20097-20106,27228-27237,... 5,1,5,...     10000
#> 10000 10600-10609,21517-21526,40461-40470,... 4,4,2,...     10001

## With a tally of the number of members in each group
reduce_DataFrame(df, df$k, count = TRUE)

#> DataFrame with 10000 rows and 8 columns
#>               k                     x                     y               z
#>       <integer>         <NumericList>         <IntegerList> <CharacterList>
#> 1             1 -0.65,-0.11,-0.53,... 36836,43803,45720,...       J,N,O,...
#> 2             2    0.06,0.16,0.13,...   6865,7895,32259,...       W,N,Y,...
#> 3             3     -1.4,0.34,1.3,...  5416,10697,16559,...       K,R,T,...
#> 4             4    0.85,0.69,0.96,...  6721,16775,19061,...       P,I,C,...
#> 5             5   -1.1,0.04,-0.29,...  7463,12833,16519,...       N,G,M,...
#> ...         ...                   ...                   ...             ...
#> 9996       9996   -1.81,0.19,0.37,...    887,9897,12215,...       O,G,I,...
#> 9997       9997   0.98,-1.66,0.07,...   3126,4474,12738,...       A,M,J,...
#> 9998       9998    -0.9,0.04,0.62,... 13651,14503,16927,...       L,B,W,...
#> 9999       9999   1.73,0.84,-0.08,...  1940,20097,27228,...       J,M,L,...
#> 10000     10000   -0.91,0.56,0.45,... 10600,21517,40461,...       X,W,I,...
#>                                            ir         r     invar        .n
#>                                 <IRangesList> <RleList> <numeric> <integer>
#> 1     36836-36845,43803-43812,45720-45729,... 3,2,3,...         2        10
#> 2         6865-6874,7895-7904,32259-32268,... 5,2,1,...         3        15
#> 3       5416-5425,10697-10706,16559-16568,... 5,2,4,...         4         8
#> 4       6721-6730,16775-16784,19061-19070,... 5,5,4,...         5        11
#> 5       7463-7472,12833-12842,16519-16528,... 1,1,1,...         6        18
#> ...                                       ...       ...       ...       ...
#> 9996        887-896,9897-9906,12215-12224,... 4,3,2,...      9997        11
#> 9997      3126-3135,4474-4483,12738-12747,... 1,4,1,...      9998        12
#> 9998  13651-13660,14503-14512,16927-16936,... 4,5,1,...      9999        14
#> 9999    1940-1949,20097-20106,27228-27237,... 5,1,5,...     10000         9
#> 10000 10600-10609,21517-21526,40461-40470,... 4,4,2,...     10001         8

## Much faster, but more crowded result
df3 <- reduce_DataFrame(df, df$k, simplify = FALSE)
df3

#> DataFrame with 10000 rows and 7 columns
#>                           k                     x                     y
#>               <IntegerList>         <NumericList>         <IntegerList>
#> 1                 1,1,1,... -0.65,-0.11,-0.53,... 36836,43803,45720,...
#> 2                 2,2,2,...    0.06,0.16,0.13,...   6865,7895,32259,...
#> 3                 3,3,3,...     -1.4,0.34,1.3,...  5416,10697,16559,...
#> 4                 4,4,4,...    0.85,0.69,0.96,...  6721,16775,19061,...
#> 5                 5,5,5,...   -1.1,0.04,-0.29,...  7463,12833,16519,...
#> ...                     ...                   ...                   ...
#> 9996     9996,9996,9996,...   -1.81,0.19,0.37,...    887,9897,12215,...
#> 9997     9997,9997,9997,...   0.98,-1.66,0.07,...   3126,4474,12738,...
#> 9998     9998,9998,9998,...    -0.9,0.04,0.62,... 13651,14503,16927,...
#> 9999     9999,9999,9999,...   1.73,0.84,-0.08,...  1940,20097,27228,...
#> 10000 10000,10000,10000,...   -0.91,0.56,0.45,... 10600,21517,40461,...
#>                     z                                      ir         r
#>       <CharacterList>                           <IRangesList> <RleList>
#> 1           J,N,O,... 36836-36845,43803-43812,45720-45729,... 3,2,3,...
#> 2           W,N,Y,...     6865-6874,7895-7904,32259-32268,... 5,2,1,...
#> 3           K,R,T,...   5416-5425,10697-10706,16559-16568,... 5,2,4,...
#> 4           P,I,C,...   6721-6730,16775-16784,19061-19070,... 5,5,4,...
#> 5           N,G,M,...   7463-7472,12833-12842,16519-16528,... 1,1,1,...
#> ...               ...                                     ...       ...
#> 9996        O,G,I,...       887-896,9897-9906,12215-12224,... 4,3,2,...
#> 9997        A,M,J,...     3126-3135,4474-4483,12738-12747,... 1,4,1,...
#> 9998        L,B,W,... 13651-13660,14503-14512,16927-16936,... 4,5,1,...
#> 9999        J,M,L,...   1940-1949,20097-20106,27228-27237,... 5,1,5,...
#> 10000       X,W,I,... 10600-10609,21517-21526,40461-40470,... 4,4,2,...
#>                       invar
#>               <NumericList>
#> 1                 2,2,2,...
#> 2                 3,3,3,...
#> 3                 4,4,4,...
#> 4                 5,5,5,...
#> 5                 6,6,6,...
#> ...                     ...
#> 9996     9997,9997,9997,...
#> 9997     9998,9998,9998,...
#> 9998     9999,9999,9999,...
#> 9999  10000,10000,10000,...
#> 10000 10001,10001,10001,...