Quantitative MS Features

Conceptually, a Features object holds a set of assays. These are composed of a matrix (or array) containing quantitative data and row annotations (meta-data). The number and the names (nature) of columns (samples) are always the same across the matrices, but the number of rows (features) can vary. Each one of these matrices has a set of feature annotations (encoded as DataFrame objects), that have the same number of rows as the assay matrix their are associated to, and an arbitrary number of columns (feature variables). Such matrix and feature annotation pairs are typically defined as MSnSet or SummarizedExperiment objects. In addition, a Features object also uses a single DataFrame to annotate the samples (columns) represented in all the matrices.

Features(...)

Arguments

...
See [MultiAssayExperiment].

Value

A new instance of class `Features`.

Details

A typical use case for such Features object is to represent quantitative proteomics (or metabolomics) data, where different assays represent quantitation data at the PSM (the main assay), peptide and protein level, and where peptide values are computed from the PSM data, and the protein-level data is calculated based on the peptide-level values.

The largest assay (the one with the highest number of features) is considered the main assay, from which the other ones are derived by aggregating/combining several rows into a single one.

The recommended way to create Features objects is the use the readFeatures() function, that creates an instance from tabular data. The Features constructor can be used to create objects from their bare parts. It is the user's responsability to make sure that these match the class validity requirements.

Accessors

See MultiAssayExperiment.

See also

Examples

## ------------------------
## An empty Features object
## ------------------------

Features()

#> A empty instance of class Features 

## -----------------------------------
## Creating a Features object manually
## -----------------------------------

## two assays (matrices) with matching column names
m1 <- matrix(1:40, ncol = 4)
m2 <- matrix(1:16, ncol = 4)
sample_names <- paste0("S", 1:4)
colnames(m1) <- colnames(m2) <- sample_names
rownames(m1) <- letters[1:10]
rownames(m2) <- letters[1:4]

## two corresponding feature metadata with appropriate row names
df1 <- DataFrame(Fa = 1:10, Fb = letters[1:10],
                 row.names = rownames(m1))
df2 <- DataFrame(row.names = rownames(m2))

(se1 <- SummarizedExperiment(m1, df1))

#> class: SummarizedExperiment 
#> dim: 10 4 
#> metadata(0):
#> assays(1): ''
#> rownames(10): a b ... i j
#> rowData names(2): Fa Fb
#> colnames(4): S1 S2 S3 S4
#> colData names(0):

(se2 <- SummarizedExperiment(m2, df2))

#> class: SummarizedExperiment 
#> dim: 4 4 
#> metadata(0):
#> assays(1): ''
#> rownames(4): a b c d
#> rowData names(0):
#> colnames(4): S1 S2 S3 S4
#> colData names(0):

## Sample annotation (colData)
cd <- DataFrame(Var1 = rnorm(4),
                Var2 = LETTERS[1:4],
                row.names = sample_names)

el <- list(assay1 = se1, assay2 = se2)
fts1 <- Features(el, colData = cd)
fts1

#> A instance of class Features containing
#>  [1] assay1: SummarizedExperiment with 10 rows and 4 columns 
#>  [2] assay2: SummarizedExperiment with 4 rows and 4 columns 

## -------------------------------------------------
## Creating a Features object from a data.frame (see
## ?readFeatures) for details
## -------------------------------------------------

data(hlpsms)
fts2 <- readFeatures(hlpsms, ecol = 1:10, name = "psms")
fts2

#> A instance of class Features containing
#>  [1] psms: SummarizedExperiment with 32776 rows and 10 columns 

fts2[[1]]

#> class: SummarizedExperiment 
#> dim: 32776 10 
#> metadata(0):
#> assays(1): ''
#> rownames(32776): 136 141 ... 115747 115775
#> rowData names(18): Sequence ProteinDescriptions ... RTmin markers
#> colnames(10): X126 X127C ... X130N X131
#> colData names(0):

fts2[["psms"]]

#> class: SummarizedExperiment 
#> dim: 32776 10 
#> metadata(0):
#> assays(1): ''
#> rownames(32776): 136 141 ... 115747 115775
#> rowData names(18): Sequence ProteinDescriptions ... RTmin markers
#> colnames(10): X126 X127C ... X130N X131
#> colData names(0):